Added Order Service

+++ bakedDelight/src/main/java/com/masai/service/CustomerServiceImpl.java
@@ -6,9 +6,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.masai.exception.CustomerException;
+import com.masai.exception.LoginException;
 import com.masai.model.Cart;
+import com.masai.model.CurrentUserSession;
 import com.masai.model.Customer;
 import com.masai.repository.CustomerRepository;
+import com.masai.repository.SessionRepository;
 
 @Service
 public class CustomerServiceImpl implements CustomerService{
@@ -16,6 +19,9 @@ public class CustomerServiceImpl implements CustomerService{
     @Autowired
     private CustomerRepository cDao;
 
+    @Autowired
+	public SessionRepository sessionRepo;
+    
     @Override
     public Customer addCustomer(Customer customer){
         
@@ -34,9 +40,22 @@ public class CustomerServiceImpl implements CustomerService{
     }
 
     @Override
-    public Customer deleteCustomer(Integer customerId) throws CustomerException {
+    public Customer deleteCustomer(Integer customerId,String key) throws CustomerException, LoginException {
         
-
+        CurrentUserSession status =	sessionRepo.findByUuid(key);
+		
+        if(status==null) {
+			
+			throw new LoginException("Customer not logged yet..plz login first");
+			
+		}
+		
+		if(status.getRole().equalsIgnoreCase("customer")) {
+			
+			throw new CustomerException("You are not authorized to add products");
+			
+		}
+        
         Customer customer = cDao.findById(customerId).orElseThrow(()-> new CustomerException("Customer Not Found..."));
 
         cDao.delete(customer);
@@ -45,15 +64,43 @@ public class CustomerServiceImpl implements CustomerService{
     }
 
     @Override
-    public Customer getCustomer(Integer customerId) throws CustomerException {
+    public Customer getCustomer(Integer customerId,String key) throws CustomerException, LoginException {
         
+        CurrentUserSession status =	sessionRepo.findByUuid(key);
+		
+        if(status==null) {
+			
+			throw new LoginException("Customer not logged yet..plz login first");
+			
+		}
+		
+		if(status.getRole().equalsIgnoreCase("customer")) {
+			
+			throw new CustomerException("You are not authorized to add products");
+			
+		}
+
         Customer customer = cDao.findById(customerId).orElseThrow(()-> new CustomerException("Customer Not Found..."));
 
         return customer;
     }
 
     @Override
-    public List<Customer> getAllCustomer() throws CustomerException {
+    public List<Customer> getAllCustomer(String key) throws CustomerException, LoginException {
+        
+        CurrentUserSession status =	sessionRepo.findByUuid(key);
+		
+        if(status==null) {
+			
+			throw new LoginException("Customer not logged yet..plz login first");
+			
+		}
+		
+		if(status.getRole().equalsIgnoreCase("customer")) {
+			
+			throw new CustomerException("You are not authorized to add products");
+			
+		}
         
         List<Customer> customers = cDao.findAll();
